# Make patch_list generating faster